# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/student-homework": {
      "post": {
        "operationId": "StudentHomeworkController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStudentHomeworkDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "StudentHomework"
        ]
      },
      "get": {
        "operationId": "StudentHomeworkController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StudentHomework"
        ]
      }
    },
    "/api/student-homework/{id}": {
      "get": {
        "operationId": "StudentHomeworkController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StudentHomework"
        ]
      },
      "patch": {
        "operationId": "StudentHomeworkController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStudentHomeworkDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StudentHomework"
        ]
      },
      "delete": {
        "operationId": "StudentHomeworkController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "StudentHomework"
        ]
      }
    },
...
    "/api/user": {
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/user/{id}": {
      "get": {
        "operationId": "UserController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "patch": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      },
      "delete": {
        "operationId": "UserController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "User"
        ]
      }
    },
    "/api/role": {
      "post": {
        "operationId": "RoleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      },
      "get": {
        "operationId": "RoleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      }
    },
    "/api/role/{id}": {
      "get": {
        "operationId": "RoleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      },
      "patch": {
        "operationId": "RoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      },
      "delete": {
        "operationId": "RoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Role"
        ]
      }
    },
    "/api/user-role": {
      "post": {
        "operationId": "UserRoleController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "UserRole"
        ]
      },
      "get": {
        "operationId": "UserRoleController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserRole"
        ]
      }
    },
    "/api/user-role/{id}": {
      "get": {
        "operationId": "UserRoleController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserRole"
        ]
      },
      "patch": {
        "operationId": "UserRoleController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserRole"
        ]
      },
      "delete": {
        "operationId": "UserRoleController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserRole"
        ]
      }
    },
    "/api/user-role/user/{userId}": {
      "get": {
        "operationId": "UserRoleController_findByUserId",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "UserRole"
        ]
      }
    },
    "/api/user-role/assign": {
      "post": {
        "operationId": "UserRoleController_assignRoles",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRolesDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "UserRole"
        ]
      }
    }
  },
  "info": {
    "title": "API Documentation",
    "description": "The API description",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateStudentHomeworkDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Description of the student homework",
            "example": "Homework description"
          },
          "student_id": {
            "type": "number",
            "description": "Student ID of the student homework",
            "example": 1
          },
          "lesson_id": {
            "type": "number",
            "description": "Lesson ID of the student homework",
            "example": 1
          }
        },
        "required": [
          "description",
          "student_id",
          "lesson_id"
        ]
      },
      "UpdateStudentHomeworkDto": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Description of the student homework",
            "example": "Homework description"
          },
          "student_id": {
            "type": "number",
            "description": "Student ID of the student homework",
            "example": 1
          },
          "lesson_id": {
            "type": "number",
            "description": "Lesson ID of the student homework",
            "example": 1
          }
        }
      },
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John",
            "description": "User first name"
          },
          "surname": {
            "type": "string",
            "example": "Doe",
            "description": "User last name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "User email"
          },
          "phone": {
            "type": "number",
            "example": 998901234567,
            "description": "User phone number"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          },
          "gender": {
            "type": "string",
            "example": "Male",
            "description": "User gender"
          },
          "xp": {
            "type": "number",
            "example": 100,
            "description": "Experience points"
          },
          "is_learning": {
            "type": "boolean",
            "example": true,
            "description": "Is learning"
          },
          "district_id": {
            "type": "number",
            "example": 1,
            "description": "District ID"
          },
          "job": {
            "type": "string",
            "example": "Teacher",
            "description": "Job title"
          },
          "hire_date": {
            "type": "string",
            "example": "2024-01-01",
            "description": "Hire date"
          },
          "salary": {
            "type": "number",
            "example": 1000,
            "description": "Salary"
          },
          "is_fired": {
            "type": "boolean",
            "example": false,
            "description": "Is fired"
          }
        },
        "required": [
          "name",
          "surname",
          "phone",
          "password"
        ]
      },
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "John",
            "description": "User first name"
          },
          "surname": {
            "type": "string",
            "example": "Doe",
            "description": "User last name"
          },
          "email": {
            "type": "string",
            "example": "john.doe@example.com",
            "description": "User email"
          },
          "phone": {
            "type": "number",
            "example": 998901234567,
            "description": "User phone number"
          },
          "password": {
            "type": "string",
            "example": "password123",
            "description": "User password"
          },
          "gender": {
            "type": "string",
            "example": "Male",
            "description": "User gender"
          },
          "xp": {
            "type": "number",
            "example": 100,
            "description": "Experience points"
          },
          "is_learning": {
            "type": "boolean",
            "example": true,
            "description": "Is learning"
          },
          "district_id": {
            "type": "number",
            "example": 1,
            "description": "District ID"
          },
          "job": {
            "type": "string",
            "example": "Teacher",
            "description": "Job title"
          },
          "hire_date": {
            "type": "string",
            "example": "2024-01-01",
            "description": "Hire date"
          },
          "salary": {
            "type": "number",
            "example": 1000,
            "description": "Salary"
          },
          "is_fired": {
            "type": "boolean",
            "example": false,
            "description": "Is fired"
          }
        }
      },
      "CreateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ADMIN",
            "description": "Role name"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateRoleDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ADMIN",
            "description": "Role name"
          }
        }
      },
      "CreateUserRoleDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "number",
            "example": 1,
            "description": "User ID to assign role to"
          },
          "role_id": {
            "type": "number",
            "example": 1,
            "description": "Role ID to assign (e.g., 1 for ADMIN, 2 for TEACHER)"
          }
        },
        "required": [
          "user_id",
          "role_id"
        ]
      },
      "UpdateUserRoleDto": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "number",
            "example": 1,
            "description": "User ID"
          },
          "role_id": {
            "type": "number",
            "example": 1,
            "description": "Role ID"
          }
        }
      },
      "AssignRolesDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "number",
            "example": 1,
            "description": "User ID to assign roles to"
          },
          "roleIds": {
            "example": [
              1,
              2
            ],
            "description": "Array of role IDs to assign (e.g., [1, 2] for both ADMIN and TEACHER roles)",
            "type": "array",
            "items": {
              "type": "number"
            }
          }
        },
        "required": [
          "userId",
          "roleIds"
        ]
      }
    }
  }
}